+++ CHANGELOG.rst
@@ -82,3 +82,4 @@ v0.12 - 2020-06-03
 Head
 ++++
 - feat: Add 'LoggedMixin' mixin class.
+- feat: Add 'debug_indent' and 'trace_indent' methods in the 'Logger' class.